+++ src/Exception/OutOfBoundsException.php
@@ -1,0 +1,31 @@
+<?php
+
+/*
+ +------------------------------------------------------------------------+
+ | Phalcon Breadcrumbs                                                    |
+ +------------------------------------------------------------------------+
+ | Copyright (c) 2016 Phalcon Framework Team                              |
+ +------------------------------------------------------------------------+
+ | This source file is subject to the New BSD License that is bundled     |
+ | with this package in the file docs/LICENSE.txt.                        |
+ |                                                                        |
+ | If you did not receive a copy of the license and are unable to         |
+ | obtain it through the world-wide-web, please send an email             |
+ | to serghei@phalconphp.com so I can send you a copy immediately.        |
+ +------------------------------------------------------------------------+
+ | Authors: Serghei Iakovlev <serghei@phalconphp.com>                     |
+ +------------------------------------------------------------------------+
+*/
+
+namespace Phalcon\Breadcrumbs\Exception;
+
+/**
+ * Phalcon\Breadcrumbs\Exception\OutOfBoundsException
+ *
+ * Exceptions thrown in Phalcon\Breadcrumbs will use this class
+ *
+ * @package Phalcon\Breadcrumbs\Exception
+ */
+class OutOfBoundsException extends \OutOfBoundsException
+{
+}