+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # [1.1.0](https://github.com/phalcongelist/breadcrumbs/releases/tag/v1.1.0) (2016-XX-XX)
 * Added support of events
 * Added `Breadcrumbs::update` to update an existing crumb
-* Added the `breadcrumbs:beforeUpdate` and `breadcrumbs:afterUpdate` events
+* Added the events: `breadcrumbs:beforeUpdate` and `breadcrumbs:afterUpdate`
+* Updated `Breadcrumbs::log` in order to add the ability to catch the exception in your custom listener
 
 # [1.0.0](https://github.com/phalcongelist/breadcrumbs/releases/tag/v1.0.0) (2016-02-21)
 * Initial release